The multiprocessing routine has been refined and improved in Counter. The new approach will handle results as soon as they become available, regardless of order of execution. This doesn't result in much of a performance increase, but it is more "correct."
FeatureCounter has also been moved to feature_selector.py.
